댓글,게시물 예외처리 완료

+++ src/main/java/com/sparta/apiserver1/Comment/service/CommentService.java
@@ -8,17 +8,22 @@ import com.sparta.apiserver1.Comment.repository.CommentRepository;
 import com.sparta.apiserver1.Post.entity.Post;
 import com.sparta.apiserver1.Post.repository.PostRepository;
 import com.sparta.apiserver1.User.entity.User;
+import com.sparta.apiserver1.User.entity.UserRoleEnum;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Locale;
+import java.util.concurrent.RejectedExecutionException;
 
 @Service
 @RequiredArgsConstructor
 public class CommentService {
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
+    private final MessageSource messageSource;
 
     public List<CommentResponseDto> selecteAllComment() {
         return commentRepository.findAllByOrderByCreatedAtDesc().stream().map(CommentResponseDto::new).toList();
@@ -38,7 +43,17 @@ public class CommentService {
     public CommentResponseDto updateComment(Long postId, Long commentId, CommentRequestDto requestDto, User user) {
         findPost(postId);
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new IllegalArgumentException("선택한 댓글이 존재하지 않습니다."));
-        comment.checkUsername(user.getUsername()); // 지금 로그인한 사람이 댓글 작성한 사람이 맞는지 확인
+        //요청자(user) 가 같은지 체크
+        if (!comment.getUsername().equals(user.getUsername())) {
+            throw new IllegalArgumentException(
+                    messageSource.getMessage(
+                            "only.writer.update",
+                            null ,
+                            "only writer can update",
+                            Locale.getDefault()
+                    )
+            );
+        }
         comment.update(requestDto);
         return new CommentResponseDto(comment);
 
@@ -46,7 +61,18 @@ public class CommentService {
     public CommentResponseDto deleteCommit(Long postId, Long commentId, User user) {
         findPost(postId);
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new IllegalArgumentException("선택한 댓글이 존재하지 않습니다."));
-        comment.checkUsername(user.getUsername());
+
+        //요청자(user) 가 같은지 체크
+        if (!comment.getUsername().equals(user.getUsername())) {
+            throw new IllegalArgumentException(
+                    messageSource.getMessage(
+                            "only.writer.delete",
+                            null ,
+                            "only writer can delete",
+                            Locale.getDefault()
+                    )
+            );
+        }
         commentRepository.delete(comment);
 
         return new CommentResponseDto(comment);
@@ -56,8 +82,6 @@ public class CommentService {
         return postRepository.findById(postId).orElseThrow(() ->
                 new IllegalArgumentException("선택한 게시물이 존재하지 않습니다"));
     }
-
-    ;
 
 
 }
